[refactor] split SearchList component
